Completed security improvement

+++ server/routes/users.js
@@ -1,11 +1,16 @@
 const User = require("../models/user");
 const express = require("express");
 const router = express.Router();
+const jwt = require("jsonwebtoken");
 
 router.post("/", async (req, res) => {
     try {
-        const user = await new User(req.body).save();
-        res.send(user);
+        const accessToken = req.headers.authorization.replace("Bearer ", "");
+        const decoded = jwt.verify(accessToken, process.env.PUBLIC_KEY, { algorithms: ["RS256"] });
+        if (decoded) {
+            const user = await new User(req.body).save();
+            res.send(user);
+        }
     } catch (error) {
         res.send(error);
     }
@@ -13,7 +18,9 @@ router.post("/", async (req, res) => {
 
 router.get("/", async (req, res) => {
     try {
-        const users = await User.find();
+        const accessToken = req.headers.authorization.replace("Bearer ", "");
+        const decoded = jwt.verify(accessToken, process.env.PUBLIC_KEY, { algorithms: ["RS256"] });
+        const users = await User.findOne({ userfrontId: decoded.userId });
         res.send(users);
     } catch (error) {
         res.send(error);
@@ -22,11 +29,15 @@ router.get("/", async (req, res) => {
 
 router.put("/:id", async (req, res) => {
     try {
-        const user = await User.findOneAndUpdate(
-            { _id: req.params.id },
-            req.body
-        );
-        res.send(user);
+        const accessToken = req.headers.authorization.replace("Bearer ", "");
+        const decoded = jwt.verify(accessToken, process.env.PUBLIC_KEY, { algorithms: ["RS256"] });
+        if (decoded) {
+            const user = await User.findOneAndUpdate(
+                { _id: req.params.id },
+                req.body
+            );
+            res.send(user);
+        }
     } catch (error) {
         res.send(error);
     }
@@ -34,8 +45,12 @@ router.put("/:id", async (req, res) => {
 
 router.delete("/:id", async (req, res) => {
     try {
-        const user = await User.findByIdAndDelete(req.params.id);
-        res.send(user);
+        const accessToken = req.headers.authorization.replace("Bearer ", "");
+        const decoded = jwt.verify(accessToken, process.env.PUBLIC_KEY, { algorithms: ["RS256"] });
+        if (decoded) {
+            const user = await User.findByIdAndDelete(req.params.id);
+            res.send(user);
+        }
     } catch (error) {
         res.send(error);
     }
